[FIX] Enable session_redis to work with Sentry
When the sentry module is loaded, it overrides odoo.http.Application
with a middleware object that wraps the original, but does not proxy it.
Thus the session_store property that session_redis monkeypatches is no
longer there.

This fix intends to avoid that, without having very specific knowledge
about the sentry module baked in.

The solution seems to work, but has the risk that if the Application
class is fully replaced (i.e. if it is not even wrapped), then this new
method may lead to this module simply not working, and the reasons for
it may be difficult to debug. But the only way around that would be to
explicitly check the wrapper object contains the expected class, and
that is basically the same as depending on the sentry module.